% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais.R
\docType{data}
\name{ais}
\alias{ais}
\title{Australia Institute of Sport data}
\description{
Data on 102 male and 100 female athletes collected at the Austrialian
Institute of Sports
}
\references{
S.Weisberg(2005). Applied Linear Regression, 3rd edition. New York,
Section 6.4.
}
\keyword{data}
\keyword{set}

