% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDqr_GCD.R
\name{ALDqr_GCD}
\alias{ALDqr_GCD}
\title{Outlier Dignostic for Quantile Regression Based on MLE Estimation}
\usage{
ALDqr_GCD(y, x, tau, error, iter)
}
\arguments{
\item{y}{dependent variable in quantile regression}

\item{x}{indepdent variables in quantile regression.
Note that: x is the independent variable matrix which including
the intercept. That means, if the dimension of independent
variables is p and the sample size is n, x is a n times p+1
matrix with the first column is one.}

\item{tau}{quantile}

\item{error}{The EM algorithm accuracy of error used in MLE estimation}

\item{iter}{the iteration frequancy for EM algorithm used in MLE estimation}
}
\description{
Outlier Dignostic for Quantile Regression Based on MLE Estimation
}

