% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesProb.R
\name{bayesProb}
\alias{bayesProb}
\title{Outlier Dignostic for Quantile Regression Based on Bayesian Estimation}
\usage{
bayesProb(y, x, tau, M)
}
\arguments{
\item{y}{dependent variable in quantile regression}

\item{x}{indepdent variables in quantile regression.
Note that: x is the independent variable matrix}

\item{tau}{quantile}

\item{M}{the iteration frequancy for MCMC used in Baysian Estimation}
}
\description{
Outlier Dignostic for Quantile Regression Based on Bayesian Estimation
}

