% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_distance_benchmark.R
\name{cook_benchmark}
\alias{cook_benchmark}
\title{Calculating the benchmark of cook distance}
\usage{
cook_benchmark(n)
}
\arguments{
\item{n}{The sample size}
}
\description{
This function calculate the benchmark of cook distance
}

