% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{es}
\alias{es}
\title{Studentised residuals based on quantile regression}
\usage{
es(object)
}
\arguments{
\item{object}{an object of class quantile regression model}
}
\description{
Function for computing studentised residuals
from a quantile regression fit.
}
