% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_bayes.R
\name{frame_bayes}
\alias{frame_bayes}
\title{Mean probability of posterior distribution  and
Kullback-Leibler divergence}
\usage{
frame_bayes(y, x, tau, M, method = c("bayes.prob", "bayes.kl"))
}
\arguments{
\item{y}{dependent variable in quantile regression}

\item{x}{indepdent variables in quantile regression.
Note that: x is the independent variable matrix with first column
being one, indicate the intercept}

\item{tau}{quantile}

\item{M}{the iteration frequancy for MCMC used in Baysian Estimation}

\item{method}{the diagnostic method for outlier detection}
}
\description{
This function give the data frame to plot  the mean probability of posterior and Kullback-leibler divergence of quantile regression model with asymmetric laplace distribution based on bayes estimation procedure.
}
\examples{
\dontrun{
library(ggplot2)
data(ais)
y <- ais$BMI
x <- ais$LBM
tau <- c(0.1, 0.5, 0.9)
case <- rep(1:length(y), length(tau))
prob <- frame_bayes(y, x, tau, M =  100,
                 method = 'bayes.prob')
prob_m <- cbind(case, prob)
ggplot(prob_m, aes(x = case, y = value )) +
  geom_point() +
  facet_wrap(~variable, scale = 'free') +
  xlab("case number") +
  ylab("Mean probability of posterior distribution")
kl <- frame_bayes(y, x, tau, M = 100,
                method = 'bayes.kl')
kl_m <- cbind(case, kl)
ggplot(kl_m, aes(x = case, y = value)) +
  geom_point() +
  facet_wrap(~variable, scale = 'free')+
  xlab('case number') +
  ylab('Kullback-Leibler')
}
}
\seealso{
qrod_bayes
}
