% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_GCD.R
\name{frame_GCD}
\alias{frame_GCD}
\title{This function used to prepare the data frame to plot general
cook distance.}
\usage{
frame_GCD(y, x, tau)
}
\arguments{
\item{tau}{}
}
\description{
This function used to prepare the data frame to plot general
cook distance.
}
\details{
General cook distance is
}
\author{
Wenjing Wang<wenjingwang1990@gmail.com>
}
