% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_mle.R
\name{frame_mle}
\alias{frame_mle}
\title{General Cook's distance or Q-function distance of quantile regression}
\usage{
frame_mle(y, x, tau, error = 1e-06, iter = 100,
  method = c("cook.distance", "qfunction"))
}
\arguments{
\item{y}{dependent variable for quantile regression}

\item{x}{design matrix for quantile regression}

\item{tau}{quantiles}

\item{error}{the EM algorithm accuracy of error used in MLE
estimation}

\item{iter}{the iteration frequancy for EM algorithm used in
estimation}

\item{method}{use method 'cook.distance' or 'qfunction'}
}
\value{
Generalized Cook's distance or Q-function distance for
multiple quantiles
}
\description{
data frame used to plot generalized cook distance or
qfunction distance
}
\details{
This function used to prepare the data frame to
plot general Cook's distance or Q-function distance
}
\examples{
library(ggplot2)
library(ALDqr)
data(ais)
y <- ais$BMI
x <- cbind(1, ais$LBM)
tau <- c(0.1, 0.5, 0.9)
case <- rep(1:length(y), length(tau))
GCD <- frame_mle(y, x, tau, error = 1e-06, iter = 100,
                 method = 'cook.distance')
GCD_m <- cbind(case, GCD)
ggplot(GCD_m, aes(x = case, y = value )) +
  geom_point() +
  facet_wrap(~variable, scale = 'free') +
  xlab("case number") +
  ylab("Generalized Cook Distance")
QD <- frame_mle(y, x, tau, error = 1e-06, iter = 100,
                method = 'qfunction')
QD_m <- cbind(case, QD)
ggplot(QD_m, aes(x = case, y = value)) +
  geom_point() +
  facet_wrap(~variable, scale = 'free')+
  xlab('case number') +
  ylab('Qfunction Distance')

}
\seealso{
see documentation of qrod_mle
}
\author{
Wenjing Wang\href{mailto:wenjingwang1990@gmail.com}{wenjingwang1990@gmail.com}
}
