% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_outliers.R
\name{multi_outliers}
\alias{multi_outliers}
\title{Procedures for the identification of multiple outliers in quantile regression}
\usage{
multi_outliers()
}
\description{
multiple outliers
}
\details{
multiple outliers
}

