% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_distance_benchmark.R
\name{q_benchmark}
\alias{q_benchmark}
\title{Calculating the benchmark of q function distance}
\usage{
q_benchmark(Hessian, q)
}
\arguments{
\item{Hessian}{the hessian matrix for benchmark calculating}

\item{q}{the parameter indicate the location of the local
influenctial observations}
}
\description{
This function calculate the benchmark of local influence assessing based
on Q function.
}
\details{
We propose the following Q-displacement function, because
of the intractable likelihood function. The Q-displacement function is:

\deqn{f_{Q}(w) = 2[Q(\hat{\theta}|\hat{\theta})-Q{\hat{\theta}(w)|\hat{\theta}}]}

where \eqn{\hat{\theta}(w)} is the estimate of \eqn{\theta} which maximizes

\deqn{Q(\theta, w|\hat{\theta})=E{L_{c}(\theta, w|Y_{c})|Y_{o}, \hat{\theta}}}

The benchmark of the q function distance is \eqn{\={M}(m_{0})+2S_{M}(m_{0})}

where,
\deqn{-2\"{Q} = \sum{\lambda_{i}e_{i}e_{i}^{'}}}
\deqn{M(m_{0}) = \sum {\~{\lambda_{i}}e_{i}^{2}}}
}
\references{
Zhu H T, Lee S Y. Local influence for incomplete data models[J]
. Journal of the Royal Statistical Society: Series B (Statistical Methodology)
, 2001, 63(1): 111-126.
}

