% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stud_resid.R
\name{stud_resid}
\alias{stud_resid}
\title{Studentised residuals based on quantile regression}
\usage{
stud_resid(object)
}
\arguments{
\item{object}{an object of class quantile regression model}
}
\description{
Function for computing studentised residuals
from a quantile regression fit.
}
\details{
The \code{stud_resid} function provides a studentized
residual on each quantile.
}
\references{
Ranganai, E(2016). On studentized residuals in the quantile
regression framework. University of South Africa.
\emph{SpringerPlus}, 5(1), 1231.
}
\author{
Wenjing Wang \email{wenjingwang1990@gmail.com}
}
