% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stud_resid_benchmark.R
\name{stud_resid_benchmark}
\alias{stud_resid_benchmark}
\title{Calculating the benchmark of Studentised residuals
of quantile regression}
\usage{
stud_resid_benchmark(alpha, p, n)
}
\description{
Benchmark of studentized residuals
}

