% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trout.R
\docType{data}
\name{trout}
\alias{trout}
\title{Trout Data}
\description{
The data set trout, which follows, includes the average
numbers of Lahontan cutthroat trout per meter of stream (Density),
and the width-to-depth ratios (WDRatio) for 71 samples.
}
\keyword{data}
\keyword{set}
