% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ais.R
\docType{data}
\name{ais}
\alias{ais}
\title{Australia Institute of Sport data}
\format{A data frame with 202 observations on the following 14 variables
\describe{
\item{Sex}{0 = male, 1 = female}
\item{Ht}{height(cm)}
\item{Wt}{weight{kg}}
\item{LBM}{lean body mass}
\item{RCC}{red cell count}
\item{WCC}{white cell count}
\item{Hc}{Hematocrit}
\item{Hg}{Hemoglobin}
\item{Ferr}{plasma ferritin concentration}
\item{BMI}{body mass index, weight/(height)^2}
\item{SSF}{sum of skin folds}
\item{Bfat}{Percent body fat}
\item{Label}{Case Lables}
\item{Sport}{Sport}
}}
\description{
Data on 102 male and 100 female athletes collected at the Austrialian
Institute of Sports
}
\references{
S.Weisberg(2005). Applied Linear Regression, 3rd edition. New York,
Section 6.4.
}
\keyword{data}
\keyword{set}
