% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALDqr_GCD.R
\name{ALDqr_GCD}
\alias{ALDqr_GCD}
\title{Calculating generalized cook distance of the MLE estimation of quantile regression
based on asymmetric laplace distribution}
\usage{
ALDqr_GCD(y, x = NLLL, tau = NULL, error = 1e-06, iter = 2000)
}
\arguments{
\item{y}{Dependent variable in quantile regression. Note that: we suppose
y follows asymmetric laplace distribution.}

\item{x}{indepdent variables in quantile regression.
Note that: x is the independent variable matrix which including
the intercept. That means, if the dimension of independent
variables is p and the sample size is n, x is a n times p+1
matrix with the first column is one.}

\item{tau}{quantile}

\item{error}{The EM algorithm accuracy of error used in MLE estimation}

\item{iter}{the iteration frequancy for EM algorithm used in MLE estimation}
}
\description{
Calculating generalized cook distance of the MLE estimation of quantile regression
based on asymmetric laplace distribution
}
\details{
Case-deletion is a classical approach to study the effects of dropping the
\eqn{i}th case from the data set. Thus, the complete-data log-likelihhod
function based on the data with \eqn{i}th cse deleted with be denoted by
\eqn{l_{c}(\theta|y_{c(i)})}. Let \eqn{\hat{\theta_{p(i)}} = (\hat{\beta^{'}_{p(i)}},
\hat{\sigam^{2}}_{(i)})^{'}} be the maximizer of the function

\deqn{Q_{(i)}(\theta|\hat{\theta})=E_{\hat{\theta}}[l_{c}(\theta|Y_{c(i)})|y]}

To assess the influence of the \eqn{i}th case on the EM estimate \eqn{\hat{\theta}},
we compare \eqn{\hat{\theta_(i)}} and \eqn{\hat{\theta}}, and if \eqn{\hat{\theta_(i)}}
is far from \eqn{\hat{\theta_(i)}} in some sense, then the \eqn{i}th case is regarded
as influential. Based on the metric for measuring the distance between \eqn{\hat{\theta_(i)}}
and \eqn{\hat{\theta}} proposed by Zhu et al.(2001), we consider here the following
generalized Cook distance:

\deqn{GD_{i} = (\hat{\theta_{(i)}}-\hat{\theta{i}})^{'}{-Q(\hat{\theta}|\hat{\theta})}
(\hat{\theta_{(i)}}-\hat{\theta{i}})}
}
\references{
Benites L E, Lachos V H, Vilca F E.(2015)``Case-Deletion
Diagnostics for Quantile Regression Using the Asymmetric Laplace
Distribution,\emph{arXiv preprint arXiv:1509.05099}.
}
\seealso{
\code{ALDqr_QD}
}

