% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesKL.R
\name{bayesKL}
\alias{bayesKL}
\title{Outlier Dignostic for Quantile Regression Based on Bayesian Estimation}
\usage{
bayesKL(y, x, tau, M)
}
\arguments{
\item{y}{dependent variable in quantile regression}

\item{x}{indepdent variables in quantile regression.
Note that: x is the independent variable matrix}

\item{tau}{quantile}

\item{M}{the iteration frequancy for MCMC used in Baysian Estimation}
}
\description{
Outlier Dignostic for Quantile Regression Based on Bayesian Estimation
}

