% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrod_mle.R
\name{qrod_mle}
\alias{qrod_mle}
\title{Outlier Dignostic for Quantile Regression with Asymmetric
Laplace Distribution}
\usage{
qrod_mle(y, x, tau, error, iter, method = c("cook.distance", "qfunction"))
}
\arguments{
\item{y}{Dependent variable in quantile regression}

\item{x}{Indepdent variables in quantile regression.
Note that: x is the independent variable matrix which including
the intercept. That means, if the dimension of independent
variables is p and the sample size is n, x is a n times p+1
matrix with the first column being one.}

\item{tau}{quantile}

\item{error}{The EM algorithm accuracy of error used in MLE
estimation}

\item{iter}{the iteration frequancy for EM algorithm used in
MLE estimation}

\item{method}{the diagnostic method for outlier detection}
}
\value{
Generalized Cook's distance or Q-function distance for
multiple quantiles
}
\description{
This function cacluate the generalized cook distance and q
function distance of quantile regression model with
asymmetric laplace distribution.
}
\details{
please refer to the reference paper
}
