% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trout.R
\docType{data}
\name{trout}
\alias{trout}
\title{Fish habbit of trout}
\format{A data frame with with 71 rows and 2 columns
\describe{
\item{wdratio}{Width-to-depth ratio of trout}
\item{density}{Numbers of Lahontan cutthroat trout per meter of stream}
}}
\description{
The data set trout, which follows, includes the average numbers of Lahontan
cutthroat trout per meter of stream, and the width-to-depth ratios for 71 samples
}
\keyword{dataset}
