% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_br.R
\name{wh}
\alias{wh}
\title{Dot plot for quantile regression fitting observations}
\usage{
wh(object, tau)
}
\arguments{
\item{object}{quantile regression model using br method}

\item{tau}{quantiles can be a single quantile or a vector of
quantiles}
}
\description{
This is a function that can be used to create dot plot for the
observations used in quantile regression fitting based on 'br'
method.
}
\examples{
data(ais)
tau <- c(0.1, 0.5, 0.9)
object1 <- rq(BMI ~ LBM, tau, method = 'br', data = ais)
frame_br(object1, tau)
ggplot(data_plot,
 aes(x=value, y=data_plot[,2])) +
 geom_point(alpha = 0.1) +
 ylab('y') +
 xlab('x') +
 facet_wrap(~variable, scales = "free_x", ncol = 2) +
 geom_point(data = choose, aes(x = value, y = y,
                                      group = tau_flag,
                                      colour = tau_flag,
                                      shape = obs)) +
 geom_line(data = choose, aes(x = value, y = y,
                                     group = tau_flag,
                                     colour = tau_flag))

object2 <- rq(BMI ~ Ht + LBM + Wt, tau, method = 'br',
            data = ais)
frame_br(object2, tau)
ggplot(data_plot,
 aes(x=value, y=data_plot[,2])) +
 geom_point(alpha = 0.1) +
 ylab('y') +
 xlab('x') +
 facet_wrap(~variable, scales = "free_x", ncol = 2) +
 geom_point(data = choose, aes(x = value, y = y,
                                      group = tau_flag,
                                      colour = tau_flag,
                                      shape = obs)) +
 geom_line(data = choose, aes(x = value, y = y,
                                     group = tau_flag,
                                     colour = tau_flag))


}
